@@ -119,31 +119,31 @@
     ]
   },
   "devDependencies": {
-    "babel-core": "^6.24.1",
+    "babel-core": "^6.26.3",
     "babel-eslint": "^7.2.3",
     "babel-jest": "^20.0.3",
-    "babel-loader": "^7.1.0",
+    "babel-loader": "^7.1.5",
     "babel-plugin-add-module-exports": "^0.2.1",
     "babel-plugin-dev-expression": "^0.2.1",
-    "babel-plugin-dynamic-import-webpack": "^1.0.1",
+    "babel-plugin-dynamic-import-webpack": "^1.1.0",
     "babel-plugin-flow-runtime": "^0.11.1",
     "babel-plugin-transform-class-properties": "^6.24.1",
     "babel-plugin-transform-es2015-classes": "^6.24.1",
-    "babel-preset-env": "^1.6.1",
+    "babel-preset-env": "^1.7.0",
     "babel-preset-react": "^6.24.1",
     "babel-preset-react-hmre": "^1.1.1",
     "babel-preset-react-optimize": "^1.0.1",
     "babel-preset-stage-0": "^6.24.1",
     "babel-register": "^6.24.1",
     "babili-webpack-plugin": "0.1.1",
-    "chalk": "^2.0.1",
-    "concurrently": "^3.5.0",
-    "cross-env": "^5.0.0",
+    "chalk": "^2.4.1",
+    "concurrently": "^3.6.1",
+    "cross-env": "^5.2.0",
     "cross-spawn": "^5.1.0",
     "css-loader": "^0.28.3",
-    "electron": "^1.6.10",
+    "electron": "^1.8.8",
     "electron-builder": "^19.8.0",
-    "electron-devtools-installer": "^2.2.0",
+    "electron-devtools-installer": "^2.2.4",
     "enzyme": "^2.9.1",
     "enzyme-to-json": "^1.5.1",
     "eslint": "^3.19.0",
@@ -151,63 +151,63 @@
     "eslint-formatter-pretty": "^1.1.0",
     "eslint-import-resolver-webpack": "^0.8.3",
     "eslint-plugin-compat": "^1.0.4",
-    "eslint-plugin-flowtype": "^2.33.0",
-    "eslint-plugin-flowtype-errors": "^3.3.0",
-    "eslint-plugin-import": "^2.6.0",
+    "eslint-plugin-flowtype": "^2.50.3",
+    "eslint-plugin-flowtype-errors": "^3.6.0",
+    "eslint-plugin-import": "^2.14.0",
     "eslint-plugin-jest": "^20.0.3",
     "eslint-plugin-jsx-a11y": "5.0.3",
-    "eslint-plugin-promise": "^3.5.0",
-    "eslint-plugin-react": "^7.1.0",
-    "express": "^4.15.3",
+    "eslint-plugin-promise": "^3.8.0",
+    "eslint-plugin-react": "^7.12.3",
+    "express": "^4.16.4",
     "extract-text-webpack-plugin": "^2.1.0",
-    "fbjs-scripts": "^0.8.0",
+    "fbjs-scripts": "^0.8.3",
     "file-loader": "^0.11.1",
     "flow-bin": "^0.48.0",
     "flow-runtime": "^0.13.0",
-    "flow-typed": "^2.1.2",
+    "flow-typed": "^2.5.1",
     "html-webpack-plugin": "^2.29.0",
     "identity-obj-proxy": "^3.0.0",
     "jest": "^20.0.4",
-    "jsdom": "^11.0.0",
+    "jsdom": "^11.12.0",
     "minimist": "^1.2.0",
-    "node-sass": "^4.5.3",
+    "node-sass": "^4.11.0",
     "react-addons-test-utils": "^15.6.0",
     "react-test-renderer": "^15.6.1",
     "redux-logger": "^3.0.6",
-    "rimraf": "^2.6.1",
+    "rimraf": "^2.6.3",
     "sass-loader": "^6.0.6",
     "sinon": "^2.3.5",
     "spectron": "^3.7.0",
     "style-loader": "^0.18.1",
-    "stylefmt": "^6.0.0",
+    "stylefmt": "^6.0.3",
     "stylelint": "^7.12.0",
     "stylelint-config-standard": "^16.0.0",
     "url-loader": "^0.5.8",
-    "webpack": "^3.0.0",
-    "webpack-bundle-analyzer": "^2.8.2",
-    "webpack-dev-server": "^2.5.0",
-    "webpack-merge": "^4.1.0"
+    "webpack": "^3.12.0",
+    "webpack-bundle-analyzer": "^2.13.1",
+    "webpack-dev-server": "^2.11.3",
+    "webpack-merge": "^4.2.1"
   },
   "dependencies": {
     "7zip-bin": "^2.2.7",
-    "classnames": "^2.2.5",
+    "classnames": "^2.2.6",
     "devtron": "^1.4.0",
     "electron-debug": "^1.2.0",
-    "electron-settings": "^3.1.1",
+    "electron-settings": "^3.2.0",
     "font-awesome": "^4.7.0",
     "fs-extra": "^5.0.0",
     "history": "^4.6.3",
-    "lodash": "^4.17.4",
-    "moment": "^2.20.1",
+    "lodash": "^4.17.11",
+    "moment": "^2.23.0",
     "react": "^15.6.1",
     "react-dom": "^15.6.1",
     "react-hot-loader": "3.0.0-beta.6",
-    "react-redux": "^5.0.5",
-    "react-router": "^4.1.1",
-    "react-router-dom": "^4.1.1",
+    "react-redux": "^5.1.1",
+    "react-router": "^4.3.1",
+    "react-router-dom": "^4.3.1",
     "react-router-redux": "^5.0.0-alpha.6",
     "redux": "^3.7.1",
-    "redux-thunk": "^2.2.0",
+    "redux-thunk": "^2.3.0",
     "semantic-ui-css": "2.2.12",
     "semantic-ui-react": "0.73.1",
     "slp-parser-js": "^2.0.2",
