@@ -1,8 +1,6 @@
-import { exec } from 'child_process';
 import { displayError } from './error';
 
-const { dialog, app } = require('electron').remote;
-const path = require('path');
+const { dialog } = require('electron').remote;
 
 export const SELECT_FOLDER = 'SELECT_FOLDER';
 export const SELECT_FILE = 'SELECT_FILE';
@@ -74,70 +72,15 @@ export function clearChanges() {
 }
 
 export function openDolphin() {
-  return (dispatch) => {
-    const platform = process.platform;
-    const isDev = process.env.NODE_ENV === "development";
-
-    const appPath = app.getAppPath();
-
-    // This is the path of dolphin after this app has been packaged
-    let dolphinPath = path.join(appPath, "../app.asar.unpacked/dolphin");
-
-    // Here we are going to build the platform-specific commands required to launch
-    // dolphin from the command line with the correct game
-    let commands, command;
-    switch (platform) {
-    case "darwin": // osx
-      // When in development mode, use the build-specific dolphin version
-      // In production mode, only the build from the correct platform should exist
-      dolphinPath = isDev ? "./app/dolphin-dev/osx" : dolphinPath;
-
-      commands = [
-        `cd "${dolphinPath}"`,
-        `open "Dolphin.app"`,
-      ];
-
-      // Join the commands with && which will execute the commands in sequence
-      command = commands.join(' && ');
-      break;
-    case "win32": // windows
-      // When in development mode, use the build-specific dolphin version
-      // In production mode, only the build from the correct platform should exist
-      dolphinPath = isDev ? "./app/dolphin-dev/windows" : dolphinPath;
-
-      commands = [
-        `cd "${dolphinPath}"`,
-        `Dolphin.exe`,
-      ];
-
-      // Join the commands with && which will execute the commands in sequence
-      command = commands.join(' && ');
-      break;
-    case "linux": // linux
-      dolphinPath = isDev ? "./app/dolphin-dev/linux" : dolphinPath;
-
-      commands = [
-        `cd "${dolphinPath}"`,
-        `./dolphin-emu`,
-      ];
-      
-      command = commands.join(' && ');
-      break;
-    default:
-      const error = displayError(
+  return (dispatch, getState) => {
+    const dolphinManager = getState().settings.dolphinManager;
+    dolphinManager.configureDolphin().catch((err) => {
+      const errorAction = displayError(
         'settings-global',
-        "The current platform is not supported"
+        err.message,
       );
-      dispatch(error);
-      break;
-    }
 
-    exec(command, (error) => {
-      // Apparently this callback happens before dolphin exits...
-      if (error) {
-        console.error(`exec error: ${error.message}`);
-        dispatch(displayError('settings-global', error.message));
-      }
+      dispatch(errorAction);
     });
   };
 }
