@@ -11,6 +11,7 @@ export default class Console extends Component {
     saveConnection: () => void,
     deleteConnection: () => void,
     connectConnection: () => void,
+    startMirroring: () => void,
     history: object,
     store: object,
   };
@@ -40,26 +41,30 @@ export default class Console extends Component {
       },
     });
   };
-  
-  onFormSubmit = (index) => {
+
+  onFormSubmit = (id) => {
     const formData = this.state.formData || {};
-    this.props.saveConnection(index, formData);
+    this.props.saveConnection(id, formData);
   };
 
   connectTo = (connection) => {
     this.props.connectConnection(connection);
   };
 
+  mirror = (connection) => {
+    this.props.startMirroring(connection);
+  }
+
   renderContent() {
     const store = this.props.store || {};
-    const connections = store.connections || [];
+    const connectionsById = store.connections || [];
 
     return (
       <Container text={true}>
         <Button color="blue" onClick={this.addConnectionClick}>
           Add Connection
         </Button>
-        {connections.map(this.renderConnection)}
+        {connectionsById.map(this.renderConnection)}
       </Container>
     );
   }
@@ -74,22 +79,25 @@ export default class Console extends Component {
             Edit
           </Button>
           {this.renderConnectButton(connection)}
+          {this.renderMirrorButton(connection)}
         </Card.Content>
       </Card>
     );
   }
 
-  renderConnectButton = (connection) => {
-    return (
-      <Button onClick={_.partial(this.connectTo, connection)}>Connect</Button>
-    );
-  }
+  renderConnectButton = (connection) => (
+    <Button onClick={_.partial(this.connectTo, connection)}>Connect</Button>
+  );
+
+  renderMirrorButton = (connection) => (
+    <Button onClick={_.partial(this.mirror, connection)}>Mirror</Button>
+  );
 
   renderEditModal() {
     const store = this.props.store || {};
-    const connectionToEdit = store.connectionToEdit;
+    const connectionToEdit = store.connectionSettingsToEdit;
 
-    const connectionIndex = _.get(connectionToEdit, 'index');
+    const connectionIndex = _.get(connectionToEdit, 'id');
     const actionText = connectionIndex === "new" ? "Add" : "Edit";
 
     return (
@@ -102,15 +110,15 @@ export default class Console extends Component {
     );
   }
 
-  renderEditForm(connection) {
-    if (!connection) {
+  renderEditForm(connectionSettings) {
+    if (!connectionSettings) {
       return null;
     }
 
     return (
-      <Form onSubmit={_.partial(this.onFormSubmit, connection.index)}>
-        <Form.Input name="ipAddress" label="IP Address" value={connection.ipAddress} onChange={this.onFieldChange} />
-        <Form.Input name="targetFolder" label="Target Folder" value={connection.targetFolder} onChange={this.onFieldChange} />
+      <Form onSubmit={_.partial(this.onFormSubmit, connectionSettings.id)}>
+        <Form.Input name="ipAddress" label="IP Address" value={connectionSettings.ipAddress} onChange={this.onFieldChange} />
+        <Form.Input name="targetFolder" label="Target Folder" value={connectionSettings.targetFolder} onChange={this.onFieldChange} />
         <Form.Button content="Submit" />
       </Form>
     );
