@@ -46,12 +46,11 @@ export default class PlayerChiclet extends Component {
   }
 
   renderPortBadge(player) {
-    const dotRadius = 1.7;
+    const dotRadius = 2;
     const centerX = 10;
     const centerY = 8;
     const spacer = 2.5;
     const color = "#919296";
-    const background = "#2D313A";
 
     let portCircles = [];
     switch (player.port) {
@@ -89,12 +88,12 @@ export default class PlayerChiclet extends Component {
     return (
       <svg width="20px" height="18px">
         <path
-          strokeWidth="1px"
+          strokeWidth="2px"
           stroke={color}
           d="M3.18,0.9375
             a9,9 0 1,0 13.5,0
             Z"
-          fill={background}
+          fillOpacity="0"
         />
         {portCircles}
       </svg>
